algs/src/main/java/com/dutianze/algs/leetcode/array/N_27_RemoveElement.java
package com.dutianze.algs.leetcode.array;

/**
 * <a href="https://leetcode.com/problems/remove-element/description/">27. Remove Element</a>
 * <h2>Easy</h2>
 * <pre>
 * Given an integer array nums and an integer val, remove all occurrences of val in nums in-place. The order of the elements may be changed. Then return the number of elements in nums which are not equal to val.
 *
 * Consider the number of elements in nums which are not equal to val be k, to get accepted, you need to do the following things:
 *
 * Change the array nums such that the first k elements of nums contain the elements which are not equal to val. The remaining elements of nums are not important as well as the size of nums.
 * Return k.
 *
 * Example 1:
 *
 * Input: nums = [3,2,2,3], val = 3
 * Output: 2, nums = [2,2,_,_]
 * Explanation: Your function should return k = 2, with the first two elements of nums being 2.
 * It does not matter what you leave beyond the returned k (hence they are underscores).
 * Example 2:
 *
 * Input: nums = [0,1,2,2,3,0,4,2], val = 2
 * Output: 5, nums = [0,1,4,0,3,_,_,_]
 * Explanation: Your function should return k = 5, with the first five elements of nums containing 0, 0, 1, 3, and 4.
 * Note that the five elements can be returned in any order.
 * It does not matter what you leave beyond the returned k (hence they are underscores).
 *
 *
 * Constraints:
 *
 * 0 <= nums.length <= 100
 * 0 <= nums[i] <= 50
 * 0 <= val <= 100
 * </pre>
 *
 * @author dutianze
 * @date 2024/1/23
 */
public class N_27_RemoveElement {

    public int removeElement(int[] nums, int val) {
        int startIndex = 0;
        int endIndex = nums.length - 1;

        while (startIndex <= endIndex) {
            if (nums[startIndex] == val) {
                nums[startIndex] = nums[endIndex];
                endIndex--;
                continue;
            }
            startIndex++;
        }
        return endIndex + 1;
    }
}
